fastapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
distro==1.9.0
docstring_parser==0.16
dotenv==0.9.9
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.0
gitdb==4.0.12
GitPython==3.1.44
groq==0.23.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
phidata==2.7.10
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
Pygments==2.19.1
python-dotenv==1.1.0
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
tomli==2.2.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
Werkzeug==3.1.3